Build the prototype of the user page that shows the need-help tasks
Create the user_profile.html file to build the skeleton of the user page. Create the user_profile_style.css file to modify the appearance of the file.
Create the user_profile_script.js file to pop up a notification box when the user clicks the delete button for a certain task. If the user click the OK button, then the task will be deleted.