Address the requested changes
Update the html file to change the icon for add-task button. Also, I use a <label> tag for the form input labels to connect the labels with input boxes. Modify the input box for rewarding points such that it can only take integer input ranging from 0 to 200. Add a new feature that the button color will change if the user click the button.
Update the css file to accomodate the changes and make the button text responsive to the screen size.