Add the feature that the added tasks are stored in datastore and each time the need-help tasks are loaded, we fetch real data from datastore instead of fake data
Create two servlet files to handle the POST and GET request for tasks. POST request is handled by TaskServlet.java file and it will store all the information of the newly created task in datastore.
MyNeedHelpTaskServlet.java file will handle the GET request for tasks whose owner is "Leonard". After that, it will create a json string and pass it to the script file. The script file will then handle the json string and create a table of tasks out of the json string.
Update the html file to add an "Abandon" button for each task in the Offer Help section and forbid the users from delete an "IN PROGRESS" task. CSS file is updated to accomodate the changes.