Implement the edit and delete task feature for the buttons on the need-help page
Update the TaskServlet.java file to add a doPut and doDelete method to handle the PUT and DELETE request sent by the user profile page. Update the html file to add a initially-hidden edit task box.
Update the script file to bind each button with its corresponding edit or delete function and implement the editTask and deleteTask functions. Create a TaskInfoServlet.java file to return all the information about a specific task.
Update the css file to accomodate the changes.