add user authentication implementation
Logged users will see their email address on the top right of the screen and a link to logout. They will also see the userpage button on the top left of the screen to access their userpage. If the user is not logged in, they will not see the go to userpage button and they will see a link to login on the top right of the screen instead.