Address the requested changes and complete the Edit task function
Create the EditTaskServlet.java file to handle the POST request for editing a stored task. Update MyNeedHelpTaskServlet.java and TaskInfoServlet.java file to add appropriate header comments for the two files.
Update the script file to remove two functions that are no longer needed. Update the TaskServlet.java file to remove the doPut method since an html form is not able to send a PUT request.
Update the html file to let the edit task form send a POST request instead of PUT request.