make requested changes
Commit renames the addOnClicks() function to the better worded addClickHandlers(), and the filterBy() function to the better suited filterTasksBy().
The helpOut(), removeTask(), and cancelHelpOut() functions now make use of the element.closest() method.
A bug was fixed that lead to guest users getting error messages as a result of the addClickHandlers() function.
In the JSP homepage file, a ternary operator was used instead to assign the categoriesClass string a value based on whether or not a user is logged in.