Address the requested changes
Update the user_profile.html file to wrap the header rows with a thead element. Script file is updated to accomodate the change. Also, modify the completeTask function such that the task will not be deleted if the user tries to mark his own task as complete.
Update the EditTaskServlet.java and TaskServlet.java file to return an error early if the input task detail is empty.