add feature that lists tasks in the homepage
Upon page load, the user's location is retrieved, converted to a neighborhood, and used to fetch and display tasks within the same neighborhood. Tasks are fetched by the TaskServlet's doGet function by filtering tasks that are within the same neighborhood (zipcode, country) and by category (if any has been specified). The servlet then constructs an HTML string with all the tasks that were returned by the prepared query and depending on whether the user is logged in and is the owner of said task the HTML will be presented differently. The response is sent back to the javascript file which then uses the response to display the tasks. Similarly, when choosing any of the categories, the same fetch request is made.