add servlet functionality to claim tasks by using the EditTaskServlet's doPost method
Commit edits task that is being claimed by the helper by updating the task's status to 'IN PROGRESS' and by updating the helper task property with the helper's user id. Thus after a helper clicks on the 'HELP OUT' button in a task in the homepage, a fetch request is sent to the EditTaskServlet that checks which edit action it's being sent (in this case a 'helpout' action), and the necessary edits are made.