Add two new features that the user can update their personal info on the user_profile page and the users' score will be automatically updated when a task is verified
Update the RetrieveUserInfo.java file to return the user's current helping points based on the user's ID.
Update the MyTaskServlet.java file to filter on the current user's ID instead of a hard coded name.
Update TaskInfoServlet.java file such that the servlet file will update the helper's score if the owner of the task verifies it.
Update the UserInfoServlet.java file to add the user's helping point as a new field of the UserInfo entity.
Update the Task.java file to accomodate the change that the input Owner and Helper are now users' ID instead of their name. The Task.java file will retrieve the user's nickname based on the ID and then set the nicknames to the private owner and helper field.
Update the user_profile.jsp file to add a new modal that allows the user to change their nickname, address and phone number.
Update the script file and css file to accomodate the newly added update info modal.