make requested changes and merge master branch changes onto current branch
A Datastore transaction was used to prevent race conditions when trying to claim open tasks in the EditTaskServlet, the getRewardingPoints method was moved to its own class and it now throws an IllegalArgumentException instead of returning -1 for illegal numbers (negative or greater than 200). Some small changes were done throughout all the servlet and jsp files to accommodate the package name change from sps to neighborgood.