Add the ability to hard-code the user's lat/long locally
This helps developers and reviewers test the site in a local context
when the URL isn't http://localhost, such as when using the Linux
container on a Chromebook. The geolocation API isn't available in these
insecure contexts, so this change allows us to bypass that and continue
testing the site.