Add the feature that when the user click on the map, a marker will be placed and the formatted address will be automatically loaded to the modal.
Update user_profile.jsp file to let the user input new country and zipcode when they edit their personal information.
Update the script file such that when the user click on the map, a request will be sent to Geocoding service and the response will tell me the formatted address, zipcode and country of the clicked position.
The formatted address, zipcode and country will be automatically loaded to the edit user info modal.