Add the feature that the country, zipcode and address of a task will be the same as the user's personal information
Update the TaskServlet.java file to fetch the current user's personal information and stored the user's country, zipcode and address with the task.
Update the UserInfoServlet.java file to store the user's zipcode and country information.
Update the account.jsp to add a placeholder for zipcode and country input.
Up
Update the user_profile.jsp to adjust the style of edit personal info modal.
Update the user_profile_script.js to automatically load the user's country and zipcode while the edit info modal is shown.
Modify the UserInfoServletTest.java to accomodate the new changes.