Address the requested changes
Update the acoount.jsp and user_profile.jsp to change the input type of zipcode, country and phone number from textarea input to text-type input.
Modify the user_profile_script.jsp to change the behavior while the user deletes a marker. Now the input for zipcode, address and country will be cleared and show a placeholder message to remind the user filling out the blank