Modify the servlet files to accomodate the change that a single task now has task detail and task overview
Update EditTaskServlet.java, TaskServlet.java, and Task.java to accomodate the change that there is a new field "Task overview".
Update the index.jsp to show a pop up window with task detail when the user click on the task name. homepage_script.js and homepage_style.css are modified to accomodate the change
user_profile_script.js is modified to automatically load the task overview when the user want to edit the task