Address Issue #84 and remove all the 400 redirects caused by empty input
Update three servlet files to remove all the 400 redirects.
Update account.jsp, index.jsp and user_profile.jsp to add onsubmit function to all the forms. The onsubmit function will check whether the input is empty or empty spaces before submitting the form. If they are empty spaces, show an alert box and mark the required areas as red.
Modify css files to accomodate the change and add the onsubmit function to the script file