Add a chat box to the task detail window of the user profile page
Create a MessageServlet.java file to handle the GET and POST request sent to /messages. If it receives a POST request, it will store the new message in the datastore along with the sender information. If it receives a GET request, it will return a list of messages related with the task.
Modify the user_profile.jsp to add the chat box and a div element to hold all the messages. Update the css file to accomodate the change.
Update the script file to add a new function loadMessages(keyString) that will add all messages returned by the response of GET request to the message container