Add a new feature that at most 10 messages are loaded each time to address the scaling issue
When there are new many messages related with a single task, loading all messages at a time will lead to serious scaling issue. Therefore, I add the feature that at most 10 messages will be loaded each time to address the issue.
Now, when the user opened the window, 10 most recent messages will be shown. If the user wants to check previous messages, they can scroll to the top and click the button to load 10 more messages.
In this way, there will always be a reasonable amount of messages displayed on the screen each time.