Update the sendMessage function to address the scaling issue
Update the sendMessage function such that the doGet function is not called every time the user send a new message. Instead, I create a div element to hold the message and add it to the end of the chat box.
In this way, I could cut down the number of unnecessary calls of the doGet function for messages so that the whole sending message process would be more efficient.