Add a new feature to delete all messages related with a task when the task is deleted, abandoned or completed
Storing all the messages sent will put great pressure on the datastore. Therefore, in order to address the scalability problem, I add this new feature that when a task is deleted, abandoned or completed, the messages related with the task is no longer needed. So we could delete these messages to save place for other messages.