Add a new feature that the user will receive a notification on the user profile page if other users have sent messages to the current user
Update the MessageServlet.java file to create a Notification datastore entity when a new message is sent.
Create the NotificationServlet.java file to handle the GET request sent from the user_profile page. When it receives a GET request, it will load all the notifications with the current user as the receiver. Then, it would count how many notifications are sent for each task. Finally, it will create a json string and send the json string back to the script.
Create the Notification.java file to define the Notification class. The class will store the task ID related with the notification, the overview of the corresponding task, and how many notifications are sent for the task.
Update the user_profile.jsp file to include a new container for the notification notice. Also, I include a modal serves as the notification box. It holds a list of anchor elements. By clicking on these anchor elements, the user can open the task detail page to see the new messages directly.
Update the user_profile_script.js file to include functions that are responsible for opening and closing the notification box. Also, add a showNotifications() function to show the notice on the main page and create the anchor elements.