Add a new feature that notifications related with a task will be automatically deleted once the user opens the task detail window and sees the new chat messages
Similar to the message entities, the amount of notification entities will increase really rapidly with all the conversations between users. Therefore, scalability issues might occur if we didn't clean up the notification entities. In this commit, I add a new doPost function to the NotificationServlet.java file.
When the user open the task detail window, they will be able to see the latest chat messages. So the notifications related with the task are no longer necessary and we are able to send POST request to the NotificationServlet. When handling the POST request, the servlet file will delete all the notifications related with the task.
In this way, we can ensure that only a reasonable amount of notifications entities are stored in the datastore.