Use Cursor to better address the scaling issue
In the past commits, I managed to load at most 10 messages at a time, but the servlet file still fetch all the message entities, which is not efficient enough. In this commit, I use the class Cursor and FetchOptions to fetch at most 10 message entities from datastore at a time. This can prevent the situation that fetching takes a lot of time when there are many message entities stored in the datastore related with the current task.