Address the scaling issues for the task table using cursors
Implement the new feature that at most 5 tasks will be shown on the screen for each task table. If there are more than 5 tasks for that table, a button will be shown. By clicking the button, 5 more tasks will be loaded.
In this way, we can prevent loading too many tasks at a time and address the scaling issue of the task entities. Also, to improve the efficiency, I reduce the number of unnecessary calls to doGet function of MyTaskServlet when the user clicks the "Load 5 more tasks" button by adding new tasks directly to the table.